+++ src/azure-cli-core/azure/cli/core/__init__.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 from __future__ import print_function
 
-__version__ = "2.0.60"
+__version__ = "2.0.62"
 
 import os
 import sys
@@ -93,7 +93,7 @@ class AzCli(CLI):
 
     def exception_handler(self, ex):  # pylint: disable=no-self-use
         from azure.cli.core.util import handle_exception
-        return handle_exception(ex, cli_ctx=self)
+        return handle_exception(ex)
 
 
 class MainCommandsLoader(CLICommandsLoader):