+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_params.py
@@ -87,7 +87,7 @@ def load_arguments(self, _):
         with self.argument_context('ad sp {}'.format(item)) as c:
             c.argument('name', name_arg_type)
             c.argument('cert', arg_group='Credential', validator=validate_cert)
-            c.argument('years', default=None, arg_group='Credential')
+            c.argument('years', type=int, default=None, arg_group='Credential')
             c.argument('end_date', default=None, arg_group='Credential',
                        help="Finer grain of expiry time if '--years' is insufficient, e.g. '2020-12-31T11:59:59+00:00' or '2299-12-31'")
             c.argument('create_cert', action='store_true', arg_group='Credential')