+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+2.3.21
+++++++
+* Minor fixes.
+
+2.3.20
+++++++
+* Minor fixes.
 
 2.3.20
 ++++++