+++ src/command_modules/azure-cli-botservice/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.1.10
+++++++
+* Minor fixes
+
 0.1.9
 +++++
 * Add `az bot prepare-deploy` to prepare for deploying bots via `az webapp`