+++ src/command_modules/azure-cli-cdn/HISTORY.rst
@@ -2,10 +2,12 @@
 
 Release History
 ===============
-
-0.3.0
+0.2.2
 +++++
 * Add support for Microsoft_Standard and Standard_ChinaCdn
+
+0.2.1
++++++
 * `cdn endpoint create/update/start/stop/delete/load/purge`: Added support for `--no-wait`.
 
 0.2.0