+++ src/command_modules/azure-cli-extension/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+0.2.5
++++++
+* Minor fixes.
+
+0.2.4
++++++
+* Minor fixes.
 
 0.2.3
 +++++