+++ src/command_modules/azure-cli-find/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+0.3.2
+++++++
+* Minor fixes
+
+0.3.1
+++++++
+* Minor fixes
+
 0.3.0
 ++++++
 * Major functionality update