+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -2,6 +2,13 @@
 
 Release History
 ===============
+2.2.14
+++++++
+* Minor fixes.
+
+2.2.13
+++++++
+* Minor fixes.
 
 2.2.12
 ++++++