+++ src/command_modules/azure-cli-kusto/HISTORY.rst
@@ -2,6 +2,9 @@
 
 Release History
 ===============
+0.2.2
++++++
+Minor fixes.
 
 0.2.1
 +++++
@@ -9,11 +12,9 @@ Release History
 
 0.2.0
 +++++
-
 * Changed hot_cache_period and soft_delete_period types for database to ISO8601 duration format.
 
 
 0.1.0
 +++++
-
 * Preview release.